Fix: Course, JdbcCourseRepository - Course final 추가 - SimpleJdbcInsert 사용

+++ src/main/java/nextstep/courses/domain/CourseRepository.java
@@ -2,7 +2,7 @@ package nextstep.courses.domain;
 
 public interface CourseRepository {
 
-    int save(Course course);
+    long save(Course course);
 
     Course findById(Long id);
 }
